+++ INTERX/database/reference.go
@@ -3,7 +3,7 @@ package database
 import (
 	"time"
 
-	interx "github.com/KiraCore/sekai/INTERX/config"
+	"github.com/KiraCore/sekai/INTERX/config"
 	"github.com/sonyarouje/simdb/db"
 )
 
@@ -23,17 +23,17 @@ func (c ReferenceData) ID() (jsonField string, value interface{}) {
 	return
 }
 
-func getReferenceDbDriver() *db.Driver {
-	driver, err := db.New(interx.GetDbCacheDir() + "ref")
-	if err != nil {
-		panic(err)
-	}
-
-	return driver
+func LoadReferenceDbDriver() {
+	driver, _ := db.New(config.GetDbCacheDir() + "/ref")
+	refDb = driver
 }
 
 // GetAllReferences is a function to get all references
 func GetAllReferences() ([]ReferenceData, error) {
+	if refDb == nil {
+		panic("cache dir not set")
+	}
+
 	DisableStdout()
 
 	var references []ReferenceData
@@ -46,6 +46,10 @@ func GetAllReferences() ([]ReferenceData, error) {
 
 // GetReference is a function to get reference by key
 func GetReference(key string) (ReferenceData, error) {
+	if refDb == nil {
+		panic("cache dir not set")
+	}
+
 	DisableStdout()
 
 	data := ReferenceData{}
@@ -58,6 +62,10 @@ func GetReference(key string) (ReferenceData, error) {
 
 // AddReference is a function to add reference
 func AddReference(key string, url string, contentLength int64, lastModified time.Time, filepath string) {
+	if refDb == nil {
+		panic("cache dir not set")
+	}
+
 	data := ReferenceData{
 		Key:           key,
 		URL:           url,
@@ -86,5 +94,5 @@ func AddReference(key string, url string, contentLength int64, lastModified time
 }
 
 var (
-	refDb *db.Driver = getReferenceDbDriver()
+	refDb *db.Driver
 )