+++ INTERX/gateway/main.go
@@ -2,24 +2,22 @@ package gateway
 
 import (
 	"context"
-	"crypto/tls"
 	"fmt"
 	"mime"
 	"net/http"
-	"os"
-	"strings"
 
-	"github.com/KiraCore/sekai/INTERX/insecure"
+	"github.com/KiraCore/sekai/INTERX/config"
+	"github.com/KiraCore/sekai/INTERX/database"
 	cosmosAuth "github.com/KiraCore/sekai/INTERX/proto-gen/cosmos/auth"
 	cosmosBank "github.com/KiraCore/sekai/INTERX/proto-gen/cosmos/bank"
 	kiraGov "github.com/KiraCore/sekai/INTERX/proto-gen/kira/gov"
-	tasks "github.com/KiraCore/sekai/INTERX/tasks"
+	kiraStaking "github.com/KiraCore/sekai/INTERX/proto-gen/kira/staking"
+	"github.com/KiraCore/sekai/INTERX/tasks"
 	"github.com/gorilla/mux"
 	"github.com/grpc-ecosystem/grpc-gateway/runtime"
 	"github.com/rakyll/statik/fs"
 	"github.com/rs/cors"
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials"
 	grpclog "google.golang.org/grpc/grpclog"
 )
 
@@ -42,9 +40,11 @@ func GetGrpcServeMux(grpcAddr string) (*runtime.ServeMux, error) {
 	// This is where the gRPC-Gateway proxies the requests.
 	// WITH_TRANSPORT_CREDENTIALS: Empty parameters mean set transport security.
 	security := grpc.WithInsecure()
-	if strings.ToLower(os.Getenv("WITH_TRANSPORT_CREDENTIALS")) == "true" {
-		security = grpc.WithTransportCredentials(credentials.NewClientTLSFromCert(insecure.CertPool, ""))
-	}
+
+	// With transport credentials
+	// if strings.ToLower(os.Getenv("WITH_TRANSPORT_CREDENTIALS")) == "true" {
+	// 	security = grpc.WithTransportCredentials(credentials.NewClientTLSFromCert(insecure.CertPool, ""))
+	// }
 
 	conn, err := grpc.DialContext(
 		context.Background(),
@@ -73,11 +73,25 @@ func GetGrpcServeMux(grpcAddr string) (*runtime.ServeMux, error) {
 		return nil, fmt.Errorf("failed to register gateway: %w", err)
 	}
 
+	err = kiraStaking.RegisterQueryHandler(context.Background(), gwCosmosmux, conn)
+	if err != nil {
+		return nil, fmt.Errorf("failed to register gateway: %w", err)
+	}
+
 	return gwCosmosmux, nil
 }
 
 // Run runs the gRPC-Gateway, dialling the provided address.
-func Run(grpcAddr string, rpcAddr string, log grpclog.LoggerV2) error {
+func Run(configFilePath string, log grpclog.LoggerV2) error {
+	config.LoadConfig(configFilePath)
+	database.LoadBlockDbDriver()
+	database.LoadFaucetDbDriver()
+	database.LoadReferenceDbDriver()
+
+	grpcAddr := config.Config.GRPC
+	rpcAddr := config.Config.RPC
+	port := config.Config.PORT
+
 	tasks.RunTasks(rpcAddr)
 
 	gwCosmosmux, err := GetGrpcServeMux(grpcAddr)
@@ -86,12 +100,6 @@ func Run(grpcAddr string, rpcAddr string, log grpclog.LoggerV2) error {
 	}
 
 	oaHander := getOpenAPIHandler()
-
-	// PORT: Empty parameters mean use port 11000.
-	port := os.Getenv("PORT")
-	if port == "" {
-		port = "11000"
-	}
 
 	router := mux.NewRouter()
 	RegisterRequest(router, gwCosmosmux, rpcAddr)
@@ -110,14 +118,14 @@ func Run(grpcAddr string, rpcAddr string, log grpclog.LoggerV2) error {
 	}
 
 	// SERVE_HTTP: Empty parameters mean use the TLS Config specified with the server.
-	if strings.ToLower(os.Getenv("SERVE_HTTP")) == "false" {
-		gwServer.TLSConfig = &tls.Config{
-			Certificates: []tls.Certificate{insecure.Cert},
-		}
+	// if strings.ToLower(os.Getenv("SERVE_HTTP")) == "false" {
+	// 	gwServer.TLSConfig = &tls.Config{
+	// 		Certificates: []tls.Certificate{insecure.Cert},
+	// 	}
 
-		log.Info("Serving gRPC-Gateway and OpenAPI Documentation on https://", gatewayAddr)
-		return fmt.Errorf("serving gRPC-Gateway server: %w", gwServer.ListenAndServeTLS("", ""))
-	}
+	// 	log.Info("Serving gRPC-Gateway and OpenAPI Documentation on https://", gatewayAddr)
+	// 	return fmt.Errorf("serving gRPC-Gateway server: %w", gwServer.ListenAndServeTLS("", ""))
+	// }
 
 	log.Info("Serving gRPC-Gateway and OpenAPI Documentation on http://", gatewayAddr)
 	return fmt.Errorf("serving gRPC-Gateway server: %w", gwServer.ListenAndServe())