+++ x/gov/keeper/keeper.go
@@ -61,6 +61,14 @@ func (k Keeper) GetNetworkProperty(ctx sdk.Context, property types.NetworkProper
 			return 1, nil
 		}
 		return 0, nil
+	case types.MischanceRankDecreaseAmount:
+		return properties.MischanceRankDecreaseAmount, nil
+	case types.InactiveRankDecreasePercent:
+		return properties.InactiveRankDecreasePercent, nil
+	case types.PoorNetworkMaxBankSend:
+		return properties.PoorNetworkMaxBankSend, nil
+	case types.MinValidators:
+		return properties.MinValidators, nil
 	default:
 		return 0, errors.New("trying to fetch network property that does not exist")
 	}
@@ -85,6 +93,14 @@ func (k Keeper) SetNetworkProperty(ctx sdk.Context, property types.NetworkProper
 			properties.EnableForeignFeePayments = true
 		}
 		properties.EnableForeignFeePayments = false
+	case types.MischanceRankDecreaseAmount:
+		properties.MischanceRankDecreaseAmount = value
+	case types.InactiveRankDecreasePercent:
+		properties.InactiveRankDecreasePercent = value
+	case types.PoorNetworkMaxBankSend:
+		properties.PoorNetworkMaxBankSend = value
+	case types.MinValidators:
+		properties.MinValidators = value
 	default:
 		return errors.New("trying to set network property that does not exist")
 	}