+++ x/gov/types/genesis.go
@@ -35,7 +35,9 @@ func DefaultGenesis() *GenesisState {
 			ProposalEnactmentTime:       2, // 2min
 			EnableForeignFeePayments:    true,
 			MischanceRankDecreaseAmount: 10,
-			InactiveRankDecreasePercent: 50, // 50%
+			InactiveRankDecreasePercent: 50,      // 50%
+			PoorNetworkMaxBankSend:      1000000, // 1M ukex
+			MinValidators:               1,
 		},
 		ExecutionFees: []*ExecutionFee{
 			{
@@ -117,6 +119,28 @@ func DefaultGenesis() *GenesisState {
 				FailureFee:        100,
 				Timeout:           10,
 				DefaultParameters: 0,
+			},
+		},
+		PoorNetworkMessages: &AllowedMessages{
+			Messages: []string{
+				kiratypes.MsgTypeProposalAssignPermission,
+				kiratypes.MsgTypeProposalSetNetworkProperty,
+				kiratypes.MsgTypeSetNetworkProperties,
+				kiratypes.MsgTypeVoteProposal,
+				kiratypes.MsgTypeClaimCouncilor,
+				kiratypes.MsgTypeWhitelistPermissions,
+				kiratypes.MsgTypeBlacklistPermissions,
+				kiratypes.MsgTypeCreateRole,
+				kiratypes.MsgTypeAssignRole,
+				kiratypes.MsgTypeRemoveRole,
+				kiratypes.MsgTypeWhitelistRolePermission,
+				kiratypes.MsgTypeBlacklistRolePermission,
+				kiratypes.MsgTypeRemoveWhitelistRolePermission,
+				kiratypes.MsgTypeRemoveBlacklistRolePermission,
+				kiratypes.MsgTypeClaimValidator,
+				kiratypes.MsgTypeActivate,
+				kiratypes.MsgTypePause,
+				kiratypes.MsgTypeUnpause,
 			},
 		},
 	}